python-backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[argon2]
python-dotenv
python-multipart
psycopg2-binary
SQLAlchemy
pgvector
boto3
langchain
langchain-openai
openai
Pillow
sentence-transformers
easyocr
